requirements.txt
pandas
numpy
faker
pyyaml
streamlit
plotly
